+++ mtgoupdater/src/mtgo_preprocessor_api.rs
@@ -1,0 +1,24 @@
+use std::process::Command;
+
+use crate::mtgoparser_bin;
+
+pub fn run_mtgo_preprocessor_example() -> Result<std::process::Output, Box<dyn std::error::Error>> {
+    let pre_processor_exec_out = Command::new(mtgoparser_bin())
+        .arg("--caller")
+        .arg("mtgoupdater")
+        .arg("--run-example")
+        .output()?;
+
+    Ok(pre_processor_exec_out)
+}
+
+pub fn run_mtgo_preprocessor_json_example(
+) -> Result<std::process::Output, Box<dyn std::error::Error>> {
+    let pre_processor_exec_out = Command::new(mtgoparser_bin())
+        .arg("--caller")
+        .arg("mtgoupdater")
+        .arg("--run-example-json")
+        .output()?;
+
+    Ok(pre_processor_exec_out)
+}
